Python03/requirements.txt
flask
pdfminer
pdfminer3k
pandas
fitz
frontend
tools
fpdf
Pillow
PyPDF2
PyMuPDF==1.17.7 
